Add comment panel to utils tab
datavisyn/reprovisyn#435
