+++ src/app/workbench/sidebar/SidebarButton.tsx
@@ -1,14 +1,25 @@
-import React from 'react';
+import React, { ComponentType } from 'react';
 
-export function SidebarButton({ isSelected, color, onClick, icon }: { isSelected: boolean; color: string; onClick: (s: string) => void; icon: string }) {
+export interface ISidebarButtonProps {
+  isSelected: boolean;
+  color: string;
+  onClick: (s: string) => void;
+  icon: string;
+  extensions?: {
+    Badge?: ComponentType;
+  };
+}
+
+export function SidebarButton({ isSelected, color, onClick, icon, extensions: { Badge } = { Badge: null } }: ISidebarButtonProps) {
   return (
     <button
-      className={`btn rounded-0 shadow-none ${isSelected ? 'btn-icon-light' : 'btn-icon-dark'}`}
+      className={`btn rounded-0 shadow-none position-relative ${isSelected ? 'btn-icon-light' : 'btn-icon-dark'}`}
       type="button"
       onClick={() => (isSelected ? onClick(null) : onClick('add'))}
       style={{ backgroundColor: isSelected ? color : 'transparent' }}
     >
       <i className={icon} />
+      {Badge ? <Badge /> : null}
     </button>
   );
 }
